+++ src/items/PageItem.php
@@ -181,6 +181,21 @@ class PageItem
                 $content['fields']['wordpressUUID']['config']['type'] = 'text';
                 $content['fields']['wordpressUUID']['config']['label'] = 'WordPress UUID';
             }
+            // Comment status
+            if (isset($pageItem->comment_status)) {
+                $content['fields']['comment-status']['value'] = $pageItem->comment_status;
+                $content['fields']['comment-status']['config']['type'] = 'text';
+                $content['fields']['comment-status']['config']['label'] = 'Comment status';
+                $content['fields']['comment-status']['config']['translatable'] = 'yes';
+            }
+            // Ping status
+            if (isset($pageItem->ping_status)) {
+                $content['fields']['ping-status']['value'] = $pageItem->ping_status;
+                $content['fields']['ping-status']['config']['type'] = 'text';
+                $content['fields']['ping-status']['config']['label'] = 'Ping status';
+                $content['fields']['ping-status']['config']['translatable'] = 'yes';
+            }
+
             $content['fields']['lang']['value'] = 'en';
 
             if ($gettingFields == 0) {
@@ -280,12 +295,15 @@ class PageItem
             $content['fields']['categories']['config']['label'] = 'categories';
         }
 
-        if (isset($pageItem->featured_media) && $pageItem->featured_media) {
-            $response = Curl::sendToRestAPI($this->_restApiAddress . '/media/' . $pageItem->featured_media);
-            $response = json_decode($response);
-            $content['fields']['featuredMedia']['value'] = $response->guid->rendered;
+        if (isset($pageItem->featured_media)) {
             $content['fields']['featuredMedia']['config']['type'] = 'file';
-            $content['fields']['featuredMedia']['config']['label'] = 'file';
+            $content['fields']['featuredMedia']['config']['label'] = 'Featured Media';
+        }
+
+        if (isset($pageItem->featured_media) && $pageItem->featured_media) {
+            $content['fields']['featuredMedia']['value'] = $pageItem->featured_media;
+        } else {
+            $content['fields']['featuredMedia']['value'] = null; 
         }
 
         if (isset($pageItem->acf) && $pageItem->acf) {